issue #7 move Relativize tests into own class
