+++ smb/src/test/java/no/uis/nio/smb/CifsTest.java
@@ -65,59 +65,6 @@ public class CifsTest extends AbstractTest {
     }
 
     @Test
-    public void relativizeSibling() throws Exception {
-        Path smbA = Paths.get(new URI("smb://localhost/public/temp/a/"));
-
-        Path smbB = Paths.get(new URI("smb://localhost/public/temp/b/"));
-
-        Path relPath = smbA.relativize(smbB);
-
-        assertThat(relPath, is(notNullValue(Path.class)));
-
-        assertThat(relPath, is(instanceOf(SMBBasePath.class)));
-
-        assertThat(relPath.toString(), is("..\\b\\"));
-    }
-
-    @Test
-    public void relativizeChild() throws Exception {
-        Path smbA = Paths.get(new URI("smb://localhost/public/temp/"));
-
-        Path smbB = Paths.get(new URI("smb://localhost/public/temp/b/"));
-
-        Path relPath = smbA.relativize(smbB);
-
-        assertThat(relPath, is(notNullValue(Path.class)));
-
-        assertThat(relPath, is(instanceOf(SMBBasePath.class)));
-
-        assertThat(relPath.toString(), is("b\\"));
-    }
-
-    @Test
-    public void relativeCousin() throws Exception {
-        Path smbA = Paths.get(new URI("smb://localhost/public/temp/a/aa/"));
-
-        Path smbB = Paths.get(new URI("smb://localhost/public/temp/b/ba/"));
-
-        Path relPath = smbA.relativize(smbB);
-
-        assertThat(relPath, is(notNullValue(Path.class)));
-
-        assertThat(relPath, is(instanceOf(SMBBasePath.class)));
-
-        assertThat(relPath.toString(), is("..\\..\\b\\ba\\"));
-
-        int count = 0;
-        for (Path path : relPath) {
-            assertThat(path, is(instanceOf(SMBBasePath.class)));
-            assertTrue(path.toString().endsWith("\\"));
-            count++;
-        }
-        assertThat(count, is(3));
-    }
-
-    @Test
     public void testUriWithSpace() throws Exception {
         URI uriSpace = createTestUri("smb", "localhost", -1, "/c$/Program Files/");
